+++ previous-versions/common/src/sdk/models/operations/getcreateupdatemodeloptionsbydatatype.ts
@@ -7,9 +7,15 @@ import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
 export class GetCreateUpdateModelOptionsByDataTypeRequest extends SpeakeasyBase {
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 
@@ -21,6 +27,9 @@ export class GetCreateUpdateModelOptionsByDataTypeRequest extends SpeakeasyBase 
 }
 
 export class GetCreateUpdateModelOptionsByDataTypeResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -36,9 +45,15 @@ export class GetCreateUpdateModelOptionsByDataTypeResponse extends SpeakeasyBase
     @SpeakeasyMetadata()
     pushOption?: shared.PushOption;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }